refactor(modeling - ordertable) : 주문테이블 패키지 구조 변경
