+++ src/main/java/kitchenpos/ordertable/domain/OrderTable.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.ordertable.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
