+++ src/main/java/kitchenpos/ordertable/ui/OrderTableRestController.java
@@ -1,7 +1,7 @@
-package kitchenpos.ui;
+package kitchenpos.ordertable.ui;
 
-import kitchenpos.application.OrderTableService;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.ordertable.application.OrderTableService;
+import kitchenpos.ordertable.domain.OrderTable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
